refactor : 이벤트 리패키징
