+++ src/social-service/src/main/java/org/palette/easelsocialservice/dto/event/details/TopicConstant.java
@@ -1,4 +1,4 @@
-package org.palette.easelsocialservice.dto.event;
+package org.palette.easelsocialservice.dto.event.details;
 
 import lombok.AllArgsConstructor;
