+++ src/social-service/src/main/java/org/palette/easelsocialservice/usecase/PaintUsecase.java
@@ -3,7 +3,7 @@ package org.palette.easelsocialservice.usecase;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.palette.easelsocialservice.dto.event.PaintCreatedEvent;
+import org.palette.easelsocialservice.dto.event.details.PaintCreatedEvent;
 import org.palette.easelsocialservice.dto.request.MentionRequest;
 import org.palette.easelsocialservice.dto.request.PaintCreateRequest;
 import org.palette.easelsocialservice.dto.request.RepaintRequest;
